Single example for useAssistant (home automation). Mark as experimental.
