+++ examples/next-openai/app/api/assistant/route.ts
@@ -1,5 +1,5 @@
 import OpenAI from 'openai';
-import { AssistantResponse } from './AssistantResponse';
+import { expertimental_AssistantResponse } from './AssistantResponse';
 import { MessageContentText } from 'openai/resources/beta/threads/messages/messages';
 
 // Create an OpenAI API client (that's edge friendly!)
@@ -10,8 +10,16 @@ const openai = new OpenAI({
 // IMPORTANT! Set the runtime to edge
 export const runtime = 'edge';
 
+const homeTemperatures = {
+  bedroom: 20,
+  'home office': 21,
+  'living room': 21,
+  kitchen: 22,
+  bathroom: 23,
+};
+
 export async function POST(req: Request) {
-  // 0. Parse the request body
+  // Parse the request body
   const input: {
     threadId: string | null;
     message: string;
@@ -20,9 +28,9 @@ export async function POST(req: Request) {
   let threadId = input.threadId;
   const message = input.message;
 
-  return AssistantResponse(
-    async ({ sendStatus, sendThreadId, sendMessage }) => {
-      // 1. Create a thread if needed
+  return expertimental_AssistantResponse(
+    async ({ sendStatus, sendThreadId, sendMessage, sendData }) => {
+      // Create a thread if needed
       if (threadId == null) {
         sendStatus({
           status: 'in_progress',
@@ -34,7 +42,7 @@ export async function POST(req: Request) {
         sendThreadId(threadId);
       }
 
-      // 2. Add a message to the thread
+      // Add a message to the thread
       sendStatus({
         status: 'in_progress',
         information: 'Adding a message to the thread...',
@@ -47,40 +55,92 @@ export async function POST(req: Request) {
         },
       );
 
-      // 3. Run the assistant on the thread
+      // Run the assistant on the thread
       sendStatus({
         status: 'in_progress',
         information: 'Running the assistant on the thread...',
       });
-      let run = await openai.beta.threads.runs.create(threadId, {
-        assistant_id: 'asst_CxNzm6Opv96Ku1PBy5yUQ9QJ',
-        instructions:
-          'Please address the user as Jane Doe. The user has a premium account.',
+      const run = await openai.beta.threads.runs.create(threadId, {
+        assistant_id: 'asst_1wdVNBWkyIHZ6TBoZTpFOZYs',
       });
 
-      // 4. Poll for status change
-      while (run.status === 'queued' || run.status === 'in_progress') {
-        // delay for 500ms:
-        await new Promise(resolve => setTimeout(resolve, 500));
+      async function waitForRun(run: OpenAI.Beta.Threads.Runs.Run) {
+        // Poll for status change
+        while (run.status === 'queued' || run.status === 'in_progress') {
+          // delay for 500ms:
+          await new Promise(resolve => setTimeout(resolve, 500));
 
-        run = await openai.beta.threads.runs.retrieve(threadId, run.id);
+          run = await openai.beta.threads.runs.retrieve(threadId!, run.id);
+        }
+
+        // Check the run status
+        if (
+          run.status === 'cancelled' ||
+          run.status === 'cancelling' ||
+          run.status === 'failed' ||
+          run.status === 'expired'
+        ) {
+          sendStatus({
+            status: 'failed',
+            information: run.status,
+          });
+          throw new Error(run.status);
+        }
+
+        if (run.status === 'requires_action') {
+          if (run.required_action?.type === 'submit_tool_outputs') {
+            // TODO support several tool calls in parallel (e.g. temperature for all rooms)
+            const toolCall =
+              run.required_action.submit_tool_outputs.tool_calls[0];
+            const parameters = JSON.parse(toolCall.function.arguments);
+
+            if (toolCall.function.name === 'getRoomTemperature') {
+              const temperature =
+                homeTemperatures[
+                  parameters.room as keyof typeof homeTemperatures
+                ];
+
+              run = await openai.beta.threads.runs.submitToolOutputs(
+                threadId!,
+                run.id,
+                {
+                  tool_outputs: [
+                    {
+                      tool_call_id: toolCall.id,
+                      output: temperature.toString(),
+                    },
+                  ],
+                },
+              );
+            } else if (toolCall.function.name === 'setRoomTemperature') {
+              const temperature = parameters.temperature;
+
+              homeTemperatures[
+                parameters.room as keyof typeof homeTemperatures
+              ] = temperature;
+
+              run = await openai.beta.threads.runs.submitToolOutputs(
+                threadId!,
+                run.id,
+                {
+                  tool_outputs: [
+                    {
+                      tool_call_id: toolCall.id,
+                      output: `New temperature ${temperature} degrees celcius set.`,
+                    },
+                  ],
+                },
+              );
+            }
+
+            await waitForRun(run);
+          }
+        }
       }
 
-      // 5. Check the run status
-      if (
-        run.status === 'cancelled' ||
-        run.status === 'cancelling' ||
-        run.status === 'failed' ||
-        run.status === 'expired'
-      ) {
-        sendStatus({
-          status: 'failed',
-          information: run.status,
-        });
-        throw new Error(run.status);
-      }
+      await waitForRun(run);
 
-      // 6. Get new thread messages (after our message)
+      // Get new thread messages (after our message)
       sendStatus({
         status: 'in_progress',
         information: 'Getting new thread messages...',
@@ -93,7 +153,7 @@ export async function POST(req: Request) {
         })
       ).data;
 
-      // 7. Send the messages
+      // Send the messages
       for (const message of responseMessages) {
         sendMessage({
           id: message.id,
